+++ packages/core/styles/src/packages/atomic/progressBar/index.ts
@@ -1,0 +1,5 @@
+import * as styles from "./nimbus-progressBar.css";
+
+export const progressBar = {
+  classnames: { ...styles },
+};