+++ packages/react/src/atomic/ProgressBar/src/index.ts
@@ -1,0 +1,7 @@
+"use client";
+
+import { ProgressBar } from "./ProgressBar";
+
+export { ProgressBar } from "./ProgressBar";
+export type { ProgressBarProps } from "./ProgressBar";
+export default ProgressBar;