+++ packages/core/styles/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 Nimbus Styles deprive all styles needed to build nimbus components.
 
-## 2025-08-01 `9.20.0`
+## 2025-08-04 `9.21.0`
 
 #### 🎉 New features
 
 - Added `fullWidth` property to the `Button` component. ([#322](https://github.com/TiendaNube/nimbus-design-system/pull/322) by [@joacotornello](https://github.com/joacotornello))
+
+## 2025-08-01 `9.20.0`
+
+#### 🎉 New features
+
+- Added `progress-bar` new atomic component styles. ([#325](https://github.com/TiendaNube/nimbus-design-system/pull/325) by [@joacotornello](https://github.com/joacotornello))
 
 ## 2025-07-30 `9.19.0`